Reorganize and document Nimbus object code
